pyproject.toml
[tool.poetry]
name = "whales-identification"
version = "0.1.0"
description = "Библиотека для идентификации морских млекопитающих"
authors = ["Sergey Vandanov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.15"
absl-py = "2.1.0"
albucore = "0.0.17"
albumentations = "1.4.18"
altair = "5.4.1"
annotated-types = "0.7.0"
appnope = "0.1.4"
asttokens = "2.4.1"
astunparse = "1.6.3"
attrs = "24.2.0"
backcall = "0.2.0"
blinker = "1.8.2"
cachetools = "5.5.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
colorama = "0.4.6"
comm = "0.2.2"
debugpy = "1.8.7"
decorator = "5.1.1"
docker-pycreds = "0.4.0"
eval-type-backport = "0.2.0"
exceptiongroup = "1.2.2"
executing = "2.1.0"
filelock = "3.16.1"
flatbuffers = "24.3.25"
fsspec = "2024.9.0"
gast = "0.6.0"
gitdb = "4.0.11"
gitpython = "3.1.43"
google-auth = "2.35.0"
google-auth-oauthlib = "1.2.1"
google-pasta = "0.2.0"
grpcio = "1.66.2"
h5py = "3.12.1"
huggingface-hub = "0.25.2"
idna = "3.10"
imageio = "2.35.1"
importlib-metadata = "8.5.0"
ipykernel = "6.29.5"
ipython = "8.28.0"
jedi = "0.19.1"
jinja2 = "3.1.4"
joblib = "1.4.2"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
jupyter-client = "8.6.3"
jupyter-core = "5.7.2"
keras = "3.6.0"
lazy-loader = "0.4"
libclang = "18.1.1"
markdown = "3.7"
markdown-it-py = "3.0.0"
markupsafe = "3.0.1"
matplotlib-inline = "0.1.7"
mdurl = "0.1.2"
ml-dtypes = "0.5.0"
mpmath = "1.3.0"
namex = "0.0.8"
narwhals = "1.9.3"
nest-asyncio = "1.6.0"
networkx = "3.4.1"
numpy = "2.1.2"
oauthlib = "3.2.2"
opencv-python = "4.10.0.84"
opencv-python-headless = "4.10.0.84"
opt-einsum = "3.4.0"
optree = "0.13.0"
packaging = "24.1"
pandas = "2.2.3"
parso = "0.8.4"
pexpect = "4.9.0"
pickleshare = "0.7.5"
pillow = "10.4.0"
platformdirs = "4.3.6"
prompt-toolkit = "3.0.48"
protobuf = "5.28.2"
psutil = "6.0.0"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pyarrow = "17.0.0"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pydeck = "0.9.1"
pygments = "2.18.0"
pympler = "1.1"
python-dateutil = "2.9.0.post0"
pytz = "2024.2"
pytz-deprecation-shim = "0.1.0.post0"
pywavelets = "1.7.0"
pyyaml = "6.0.2"
pyzmq = "26.2.0"
qudida = "0.0.4"
referencing = "0.35.1"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rich = "13.9.2"
rpds-py = "0.20.0"
rsa = "4.9"
safetensors = "0.4.5"
scikit-image = "0.24.0"
scikit-learn = "1.5.2"
scipy = "1.14.1"
sentry-sdk = "2.16.0"
setproctitle = "1.3.3"
six = "1.16.0"
smmap = "5.0.1"
stack-data = "0.6.3"
streamlit = "1.39.0"
sympy = "1.13.3"
tenacity = "9.0.0"
termcolor = "2.5.0"
threadpoolctl = "3.5.0"
tifffile = "2024.9.20"
timm = "1.0.9"
toml = "0.10.2"
toolz = "1.0.0"
torch = "2.4.1"
torchvision = "0.19.1"
tornado = "6.4.1"
tqdm = "4.66.5"
traitlets = "5.14.3"
typing-extensions = "4.12.2"
tzdata = "2024.2"
tzlocal = "5.2"
urllib3 = "2.2.3"
validators = "0.34.0"
wandb = "0.18.3"
watchdog = "5.0.3"
wcwidth = "0.2.13"
werkzeug = "3.0.4"
wrapt = "1.16.0"
zipp = "3.20.2"
pytest = "^8.3.3"
pylint = "^3.3.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
